myproject/pyproject.toml
[tool.poetry]
name = "ecommerce"
version = "0.1.2"
description = "shoping"
authors = ["Your Name <you@example.com>"]
license = "n"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
django = "^5.1.1"
djangorestframework = "^3.15.2"
drf-yasg = "^1.21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
